Added onError() to handle Exceptions and Wrote few test cases

+++ src/main/java/org/arjun/cataloguemicroservice/server/CatalogueService.java
@@ -1,6 +1,7 @@
 package org.arjun.cataloguemicroservice.server;
 
 import com.google.protobuf.Empty;
+import io.grpc.Status;
 import io.grpc.stub.StreamObserver;
 import java.util.Optional;
 import org.arjun.cataloguemicroservice.*;
@@ -21,6 +22,13 @@ public class CatalogueService extends CatalogueServiceGrpc.CatalogueServiceImplB
   @Override
   public void createCatalogue(final CreateCatalogueRequest request,
                               final StreamObserver<Catalogue> responseObserver) {
+    if (catalogueServiceUtil.checkCatalogueExistenceByUserIdAndCatalogueName(
+            request.getCatalogue().getUserId(),request.getCatalogue().getCatalogueName())) {
+      Status status = Status.ALREADY_EXISTS
+              .withDescription("Catalogue with the given userId and Catalogue Name already present");
+      responseObserver.onError(status.asRuntimeException());
+      return;
+    }
     responseObserver.onNext(catalogueServiceUtil.toProto(catalogueServiceUtil
             .createCatalogueService(converter.toCatalogue(request))));
     responseObserver.onCompleted();
@@ -29,19 +37,33 @@ public class CatalogueService extends CatalogueServiceGrpc.CatalogueServiceImplB
   @Override
   public void deleteCatalogue(final DeleteCatalogueRequest request,
                               final StreamObserver<Empty> responseObserver) {
-    catalogueServiceUtil.deleteCatalogueService(request.getCatalogueId());
-    responseObserver.onNext(Empty.getDefaultInstance());
-    responseObserver.onCompleted();
+    if (catalogueServiceUtil.checkCatalogueExistenceById(request.getCatalogueId())) {
+      catalogueServiceUtil.deleteCatalogueService(request.getCatalogueId());
+      responseObserver.onNext(Empty.getDefaultInstance());
+      responseObserver.onCompleted();
+    } else {
+      Status status = Status.NOT_FOUND
+              .withDescription("Catalogue with the given Catalogue Id not found");
+      responseObserver.onError(status.asRuntimeException());
+      return;
+    }
   }
 
   @Override
   public void getCatalogue(final GetCatalogueRequest request,
                            final StreamObserver<Catalogue> responseObserver) {
-    final Optional<org.arjun.cataloguemicroservice.entity.Catalogue> entry =
-            catalogueServiceUtil.getCatalogueService(request.getCatalogueId());
-    entry.map(e -> catalogueServiceUtil.toProto(e))
-            .ifPresent(responseObserver::onNext);
-    responseObserver.onCompleted();
+    if (catalogueServiceUtil.checkCatalogueExistenceById(request.getCatalogueId())) {
+      final Optional<org.arjun.cataloguemicroservice.entity.Catalogue> entry =
+              catalogueServiceUtil.getCatalogueService(request.getCatalogueId());
+      entry.map(e -> catalogueServiceUtil.toProto(e))
+              .ifPresent(responseObserver::onNext);
+      responseObserver.onCompleted();
+    } else {
+      Status status = Status.NOT_FOUND
+              .withDescription("Catalogue with the given Catalogue Id not found");
+      responseObserver.onError(status.asRuntimeException());
+      return;
+    }
   }
 
   @Override
@@ -52,14 +74,13 @@ public class CatalogueService extends CatalogueServiceGrpc.CatalogueServiceImplB
         responseObserver.onNext(CatalogueStream.newBuilder()
                 .setCatalogue(catalogueServiceUtil.toProto(ele)).build());
       });
-      responseObserver.onCompleted();
     } else {
       catalogueServiceUtil.getCatalogueStreamByUserId(request.getUserId())
               .forEach(ele -> {
                 responseObserver.onNext(CatalogueStream.newBuilder()
                         .setCatalogue(catalogueServiceUtil.toProto(ele)).build());
               });
-      responseObserver.onCompleted();
     }
+    responseObserver.onCompleted();
   }
 }
