+++ src/main/java/org/arjun/cataloguemicroservice/server/UserService.java
@@ -1,6 +1,7 @@
 package org.arjun.cataloguemicroservice.server;
 
 import com.google.protobuf.Empty;
+import io.grpc.Status;
 import io.grpc.stub.StreamObserver;
 import java.util.Optional;
 import org.arjun.cataloguemicroservice.*;
@@ -22,6 +23,12 @@ public class UserService extends UserServiceGrpc.UserServiceImplBase {
   @Override
   public void createUser(final CreateUserRequest request,
                          final StreamObserver<User> responseObserver) {
+    if (userServiceUtil.checkUserExistenceByUserName(request.getUser().getUsername())) {
+      Status status = Status.ALREADY_EXISTS
+              .withDescription("User with the given User Name already present");
+      responseObserver.onError(status.asRuntimeException());
+      return;
+    }
     responseObserver.onNext(userServiceUtil.toProto(userServiceUtil.createUserService(
             converter.toUser(request))));
     responseObserver.onCompleted();
@@ -30,18 +37,32 @@ public class UserService extends UserServiceGrpc.UserServiceImplBase {
   @Override
   public void deleteUser(final DeleteUserRequest request, final StreamObserver<Empty>
           responseObserver) {
-    userServiceUtil.deleteUserService(request.getUserId());
-    responseObserver.onNext(Empty.getDefaultInstance());
-    responseObserver.onCompleted();
+    if (userServiceUtil.checkUserExistenceById(request.getUserId())) {
+      userServiceUtil.deleteUserService(request.getUserId());
+      responseObserver.onNext(Empty.getDefaultInstance());
+      responseObserver.onCompleted();
+    } else {
+      Status status = Status.NOT_FOUND
+              .withDescription("User with the given User Id not found");
+      responseObserver.onError(status.asRuntimeException());
+      return;
+    }
   }
 
   @Override
   public void getUser(final GetUserRequest request, final StreamObserver<User> responseObserver) {
-    final Optional<org.arjun.cataloguemicroservice.entity.User> entry =
-            userServiceUtil.getUserService(request.getUserId());
-    entry.map(e -> userServiceUtil.toProto(e))
-            .ifPresent(responseObserver::onNext);
-    responseObserver.onCompleted();
+    if (userServiceUtil.checkUserExistenceById(request.getUserId())) {
+      final Optional<org.arjun.cataloguemicroservice.entity.User> entry =
+              userServiceUtil.getUserService(request.getUserId());
+      entry.map(e -> userServiceUtil.toProto(e))
+              .ifPresent(responseObserver::onNext);
+      responseObserver.onCompleted();
+    } else {
+      Status status = Status.NOT_FOUND
+              .withDescription("User with the given User Id not found");
+      responseObserver.onError(status.asRuntimeException());
+      return;
+    }
   }
 
   @Override
